requirements.txt
streamlit==1.46.0
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
openai==1.88.0
chromadb==1.0.13
python-dotenv==1.0.0
tiktoken==0.9.0
pypdf==5.6.0
PyPDF2==3.0.1
sentence-transformers==4.1.0
numpy==2.1.3